Fixed minor issue

+++ Cesium.Runtime/FuncPtr.cs
@@ -5,14 +5,14 @@ namespace Cesium.Runtime;
 /// <summary>A class encapsulating a C function pointer.</summary>
 public readonly unsafe struct FuncPtr<TDelegate> where TDelegate : MulticastDelegate // TODO[#487]: Think about vararg and empty parameter list encoding.
 {
-    private readonly IntPtr _value;
+    private readonly long _value;
 
     public FuncPtr(void* ptr)
     {
-        _value = (IntPtr)ptr;
+        _value = (long)ptr;
     }
 
-    public static implicit operator TDelegate(FuncPtr<TDelegate> funcPtr) => (TDelegate)Activator.CreateInstance(typeof(TDelegate), [null, funcPtr._value])!;
+    public static implicit operator TDelegate(FuncPtr<TDelegate> funcPtr) => (TDelegate)Activator.CreateInstance(typeof(TDelegate), [null, (IntPtr)funcPtr._value])!;
     public static implicit operator FuncPtr<TDelegate>(TDelegate @delegate) => @delegate.Method.MethodHandle.GetFunctionPointer();
     public static implicit operator FuncPtr<TDelegate>(IntPtr funcPtr) => new((void*)funcPtr);
     public static implicit operator FuncPtr<TDelegate>(void* funcPtr) => new(funcPtr);
